Refactor & Enums
- Added UserNotificationSettings and GuildNotificationSettings enums, their order corresponding to the one in discord-bot-commands.md.
- Created an exception package and refactored exceptions to their respective packages.
- Added function to UserService that will create or re-link a BotUser, intended to be used by /link
