+++ src/main/java/com/lelebees/imperabot/bot/domain/game/exception/GameNotFoundException.java
@@ -1,4 +1,4 @@
-package com.lelebees.imperabot.bot.domain.game;
+package com.lelebees.imperabot.bot.domain.game.exception;
 
 public class GameNotFoundException extends RuntimeException {
     public GameNotFoundException(String message) {
