+++ src/main/java/com/lelebees/imperabot/bot/domain/guild/GuildNotificationSettings.java
@@ -1,0 +1,6 @@
+package com.lelebees.imperabot.bot.domain.guild;
+
+public enum GuildNotificationSettings {
+    NO_NOTIFICATIONS,
+    NOTIFICATIONS_ON
+}
