+++ src/main/java/com/lelebees/imperabot/bot/domain/user/UserNotificationSettings.java
@@ -1,0 +1,9 @@
+package com.lelebees.imperabot.bot.domain.user;
+
+public enum UserNotificationSettings {
+    NO_NOTIFICATIONS,
+    DMS_ONLY,
+    GUILD_ONLY,
+    PREFER_GUILD_OVER_DMS,
+    DMS_AND_GUILD
+}
