+++ src/main/java/com/lelebees/imperabot/bot/domain/user/exception/UserAlreadyVerfiedException.java
@@ -1,4 +1,4 @@
-package com.lelebees.imperabot.bot.domain.user;
+package com.lelebees.imperabot.bot.domain.user.exception;
 
 public class UserAlreadyVerfiedException extends RuntimeException {
     public UserAlreadyVerfiedException(String message) {
