+++ src/main/java/com/lelebees/imperabot/bot/domain/user/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.lelebees.imperabot.bot.domain.user;
+package com.lelebees.imperabot.bot.domain.user.exception;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
